test/: Reintroduce long long printf format specifiers.
We disabled the use of %lld in commit 25815be1 ("test/: Avoid long
long printf format specifiers."), presumably due to them not being
supported on some older platforms that ivykis supports.

However, these tests all run for 5 seconds by default, which is more
than 2^32 nanoseconds, which means that we will always truncate these
fields on 32 bit platforms when printing the test results.

Some googling suggests that the %lld format specifier is part of C99,
and getting the test output right on modern 32 bit platforms seems
more important than not getting the right output on some ancient
platform that doesn't implement a 20+ year old standard -- therefore,
let's revert this change.

+++ test/iv_event_raw_bench.c
@@ -93,8 +93,8 @@ int main()
 	nsec = 1000000000ULL * (tim_end.tv_sec - tim_start.tv_sec) +
 		(tim_end.tv_nsec - tim_start.tv_nsec);
 
-	printf("%s: %d in %ld nsec => %d/sec\n",
-	       iv_poll_method_name(), ev_received, (long)nsec,
+	printf("%s: %d in %lld nsec => %d/sec\n",
+	       iv_poll_method_name(), ev_received, nsec,
 	       (int)(1000000000ULL * ev_received / nsec));
 
 	return 0;
